Add author information handling in ContentEditPage; update visualizer data passing and redirect logic in TemplateManager
