+++ src/app/application/[_teamid]/templates/edit/[_templateid]/templateEdit.tsx
@@ -67,6 +67,9 @@ export default function TemplateManager({ params }: { params: { _teamid: any; _t
   const searchParams = useSearchParams();
   const getPage = useQuery(api.page.getPage, { _id: teamid as any })
   const getTemplates = useQuery(api.template.getExactTemplate, { pageid: teamid, _id: templateid })
+  if (!getTemplates) {
+    window.location.href = `/application/${teamid}/templates`
+  }
   const getFields = useQuery(api.template.getFields, { templateid: templateid })
   const templateaddfield = useMutation(api.template.addField)
   const getRole = useQuery(api.page.getRoledetail, { externalId: userId || "none", pageId: teamid });
@@ -134,6 +137,7 @@ export default function TemplateManager({ params }: { params: { _teamid: any; _t
           await onDelete(id)
           setSuccess(true)
           setTimeout(() => setIsOpen(false), 2000) // Close dialog after 2 seconds
+          window.location.href = `/application/${teamid}/templates`
         } catch (err) {
           setError("An error occurred while deleting the template. Please try again.")
         } finally {
